+++ tests/test_04_online.py
@@ -42,7 +42,10 @@ def test_connect_ftp_link():
 
 
 def test_read_url(
-    url="http://ftp.xenbase.org/pub/Genomics/JGI/README", expected="The data"
+    url="https://mbdata.science.ru.nl/siebrenf/GRCh37/GRCh37/README.txt",
+    expected="name: GRCh37",
+    # TODO: uncomment when xenbase's FTP is up again
+    # url="http://ftp.xenbase.org/pub/Genomics/JGI/README", expected="The data"
 ):
     text = genomepy.online.read_url(url)
     assert text.startswith(expected)
@@ -64,7 +67,11 @@ def test_retry(capsys):
 
 def test_check_url():
     # good URL
-    assert genomepy.online.check_url("http://ftp.xenbase.org/pub/Genomics/JGI/README")
+    assert genomepy.online.check_url(
+        "https://mbdata.science.ru.nl/siebrenf/GRCh37/GRCh37/README.txt"
+    )
+    # TODO: uncomment when xenbase's FTP is up again
+    # assert genomepy.online.check_url("http://ftp.xenbase.org/pub/Genomics/JGI/README")
 
     # bad URL:
     assert not genomepy.online.check_url("https://www.thiswebsiteisoffline.nl/")